Parsing of the selections
The parser uses a standard shunting-yard algorithm, with a higher
precedence for the `== != < <= > >=` operators, in order to minimize the
number of perentheses. All the operators are left-associative, but
parentheses should be used to deambiguate selections like

    name foo and index 2 or name bar

Implemented expressions contains `and`, `or`, `not`, `name`, `index`,
`x|y|z`, `vx|vy|vz`.