Lower the optimization level for intel compiler
It is optimizing the loop too much, and if Topology::bonds is
{{0, 1}, {2, 3}, {4, 5}}, the C API would see {{0, 1}, {0, 1}, {4, 5}}
in the data parameter.

Inserting a call with side effect (std::cout << std::endl) prevent this
from hapening, as well as calling std::memcpy directly. Because
std::memcpy can not be used here (sizeof(size_t) != sizeof(uint64_t) in
32-bits systems), we cap the optimization level for the intel compiler.