Test all pairs (triplets, quadruplets) in selection
Previous code was only testing pairs (i, j) with i < j. A selection like "pairs: name(#1) O and name(#2) H"
would not give all the matching pairs, depending on the order of the atoms in the frame.

This also change the code to test for potential matches one by one, because generating all the
potential matches would allocate a lot of memory and overflow when a frame would contain a few hundred atoms.

The new code should allocate less memory, at the price of doing more virtual calls.