Do not set C standard for external code
This was at the origin of the segfaults on OpenSuse build service:
strdup was undefined when compiling C code with -std=c99, and got a
default prototype of `int strdup(char*)`. From there, assinging the
result of strdup to a char* variable caused the pointer to be truncated
to 32bit on 64 bit platforms.

As usual with C, the fix is simple once you get everything, but understading
what is going on is hard ...