Fix a bug in shared_alloc
The pointers_ map was not updated as soon as a pointer was released, but only after
all pointers sharing the same metadata block were released. But shared pointers
could become invalid (for example by adding more atoms to a frame) and re-used by
the system allocator. These pointer would then be re-used by the system allocator,
and reach the "pointer already managed by shared_allocator" error.

At the same time, a given shared pointer can be registered multiple time
with shared_allocator (accessing the same atom twice). We don't want
to remove the corresponding pointer from the map on the first call to
free, but rather on the last. To solve this, this commit uses a
std::unordered_multimap instead of a std::unordered_map. This means that
each call to shared_ptr will add a new entry into the map, and they will
be removed one by one in free.